Move component displayTablePosition to the common component for split position
